+++ CorporateQNA/Migrations/ContextModelSnapshot.cs
@@ -117,9 +117,9 @@ namespace CorporateQNA.Migrations
                     b.Property<bool>("TwoFactorEnabled")
                         .HasColumnType("bit");
 
-                    b.Property<int>("UserId")
+                    b.Property<long>("UserId")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int")
+                        .HasColumnType("bigint")
                         .UseIdentityColumn(100);
 
                     b.Property<string>("UserName")
